+++ src/polygonParts/models/polygonPartsManager.ts
@@ -6,11 +6,11 @@ import { inject, injectable } from 'tsyringe';
 import type { EntityManager, SelectQueryBuilder } from 'typeorm';
 import { ConnectionManager } from '../../common/connectionManager';
 import { SERVICES } from '../../common/constants';
+import { ValidationError } from '../../common/errors';
 import type { ApplicationConfig, DbConfig, IConfig } from '../../common/interfaces';
 import { Part } from '../DAL/part';
 import { PolygonPart } from '../DAL/polygonPart';
-import { payloadToInsertPartsData } from '../DAL/utils';
-import { ValidationError } from '../../common/errors';
+import { payloadToInsertPartsData, setRepositoryTablePath } from '../DAL/utils';
 import {
   findSelectOutputColumns,
   geometryColumn,
@@ -22,14 +22,17 @@ import {
 } from './constants';
 import type {
   AggregateLayerMetadataOptions,
+  AggregationLayerMetadataResponse,
+  DatabaseObjectQualifiedName,
   EntitiesMetadata,
   EntityName,
   EntityNames,
+  ExistsOptions,
+  ExistsResponse,
   FilterQueryMetadata,
   FindPolygonPartsOptions,
   FindPolygonPartsQueryResponse,
   FindPolygonPartsResponse,
-  AggregationLayerMetadataResponse,
   FindQueryFilterOptions,
   FindQuerySelectOptions,
   IsValidDetailsResult,
@@ -67,7 +70,10 @@ export class PolygonPartsManager {
         };
 
         await entityManager.query(`SET search_path TO ${this.schema},public`);
-        await this.verifyAvailableTableNames(baseIngestionContext);
+        const existingEntities = await this.checkExistingEntities(baseIngestionContext);
+        if (existingEntities.length > 0) {
+          throw new ConflictError(`Table(s) with the name '${existingEntities.join()}' already exist(s)`);
+        }
         const ingestionContext = { ...baseIngestionContext, polygonPartsPayload };
         await this.createTables(ingestionContext);
         await this.insertParts(ingestionContext);
@@ -80,6 +86,37 @@ export class PolygonPartsManager {
     } catch (error) {
       const errorMessage = 'Create polygon parts transaction failed';
       logger.error({ msg: errorMessage, error });
+      throw error;
+    }
+  }
+
+  public async existsPolygonParts(options: ExistsOptions): Promise<ExistsResponse> {
+    const { entitiesMetadata, payload } = options;
+    this.logger.info({ msg: 'Checking polygon parts exists', payload });
+
+    try {
+      const polygonPartsEntityName = await this.connectionManager.getDataSource().transaction(async (entityManager) => {
+        const baseIngestionContext = {
+          entityManager,
+          logger: this.logger,
+          entitiesMetadata,
+        };
+
+        await entityManager.query(`SET search_path TO ${this.schema},public`);
+        const existingEntities = await this.checkExistingEntities(baseIngestionContext);
+        if (existingEntities.length === 0) {
+          throw new NotFoundError('Entities do not exist(s)');
+        } else if (existingEntities.length !== Object.keys(entitiesMetadata.entitiesNames).length) {
+          throw new InternalServerError('Some entities are missing');
+        }
+
+        return entitiesMetadata.entityIdentifier;
+      });
+
+      return { polygonPartsEntityName };
+    } catch (error) {
+      const errorMessage = 'Cheking polygon parts exists transaction failed';
+      this.logger.error({ msg: errorMessage, error });
       throw error;
     }
   }
@@ -283,9 +320,11 @@ export class PolygonPartsManager {
     const baseTable = filterQueryMetadata?.filterQueryAlias ?? polygonPartsEntityName.databaseObjectQualifiedName;
     const queryBuilder = filteredPolygonPartsQuery ?? entityManager.createQueryBuilder();
 
+    // eslint-disable-next-line @typescript-eslint/no-magic-numbers
+    const precision = Math.pow(10, -maxDecimalDigits);
     const footprintUnionCTE = entityManager
       .createQueryBuilder()
-      .select('st_union("polygon_part".footprint)', 'footprint_union')
+      .select(`st_union(st_reduceprecision("polygon_part".footprint, ${precision}), ${precision})`, 'footprint_union')
       .from(baseTable, 'polygon_part');
 
     const footprintSmoothCTE = entityManager
@@ -388,22 +427,20 @@ export class PolygonPartsManager {
     return aggregationQueryBuilder;
   }
 
-  private async verifyAvailableTableNames(context: {
+  private async checkExistingEntities(context: {
     entityManager: EntityManager;
     logger: Logger;
     entitiesMetadata: EntitiesMetadata;
-  }): Promise<void> {
+  }): Promise<DatabaseObjectQualifiedName[]> {
     const { entityManager, logger, entitiesMetadata } = context;
     const { entitiesNames } = entitiesMetadata;
-    logger.debug({ msg: 'Verifying polygon parts table names are available' });
+    logger.debug({ msg: 'Checking polygon parts entities existence', entitiesMetadata });
 
-    await Promise.all(
-      Object.values<EntityNames>({ ...entitiesNames }).map(async ({ databaseObjectQualifiedName, entityName }) => {
+    const entitiesExist = await Promise.all<[DatabaseObjectQualifiedName, boolean]>(
+      Object.values(entitiesNames).map(async ({ databaseObjectQualifiedName, entityName }) => {
         try {
-          const exists = await this.connectionManager.entityExists(entityManager, entityName);
-          if (exists) {
-            throw new ConflictError(`Table with the name '${databaseObjectQualifiedName}' already exists`);
-          }
+          const entityExists = await this.connectionManager.entityExists(entityManager, entityName);
+          return [databaseObjectQualifiedName, entityExists] satisfies [DatabaseObjectQualifiedName, boolean];
         } catch (error) {
           const errorMessage = `Could not verify polygon parts table name '${databaseObjectQualifiedName}' is available`;
           logger.error({ msg: errorMessage, error });
@@ -411,6 +448,7 @@ export class PolygonPartsManager {
         }
       })
     );
+    return entitiesExist.filter(([, exists]) => exists).map(([databaseObjectQualifiedName]) => databaseObjectQualifiedName);
   }
 
   private buildFindQuery<ShouldClip extends boolean = boolean>(
@@ -485,7 +523,7 @@ export class PolygonPartsManager {
       shouldClip,
     } = context;
     const polygonPart = entityManager.getRepository(PolygonPart);
-    polygonPart.metadata.tablePath = polygonPartsEntityName.databaseObjectQualifiedName; // this approach may be unstable for other versions of typeorm - https://github.com/typeorm/typeorm/issues/4245#issuecomment-2134156283
+    setRepositoryTablePath(polygonPart, polygonPartsEntityName.databaseObjectQualifiedName);
 
     const inputFilterGeometriesCTE = `select *
       from jsonb_to_recordset('${JSON.stringify(inputFilter)}'::jsonb -> 'features') as x(geometry jsonb, properties jsonb, id jsonb)`;
@@ -641,7 +679,7 @@ export class PolygonPartsManager {
 
     try {
       const part = entityManager.getRepository(Part);
-      part.metadata.tablePath = partsEntityQualifiedName; // this approach may be unstable for other versions of typeorm - https://github.com/typeorm/typeorm/issues/4245#issuecomment-2134156283
+      setRepositoryTablePath(part, partsEntityQualifiedName);
       await part.save(insertPartsData, { chunk: this.applicationConfig.chunkSize });
     } catch (error) {
       const errorMessage = `Could not insert polygon parts data to table '${partsEntityQualifiedName}'`;