+++ tests/integration/polygonParts/helpers/db.ts
@@ -6,6 +6,8 @@ import type { Feature, Polygon } from 'geojson';
 import { randexp } from 'randexp';
 import { DataSource, type DataSourceOptions, type EntityTarget, type ObjectLiteral } from 'typeorm';
 import { DatabaseCreateContext, createDatabase, dropDatabase } from 'typeorm-extension';
+import { setRepositoryTablePath } from '../../../../src/polygonParts/DAL/utils';
+import type { ExistsRequestBody } from '../../../../src/polygonParts/controllers/interfaces';
 import type { PolygonPartsPayload } from '../../../../src/polygonParts/models/interfaces';
 import type { DeepPartial } from './types';
 
@@ -75,7 +77,13 @@ export const generatePolygonPart = (): PolygonPart => {
   };
 };
 
-// TODO: merge with generateRequest() in requestsMocks.ts
+export const generateExistsPayload = (): ExistsRequestBody => {
+  return {
+    productId: generateProductId(),
+    productType: generateProductType(),
+  };
+};
+
 export function generatePolygonPartsPayload(partsCount: number): PolygonPartsPayload;
 export function generatePolygonPartsPayload(template: DeepPartial<PolygonPartsPayload>): PolygonPartsPayload;
 export function generatePolygonPartsPayload(input: number | DeepPartial<PolygonPartsPayload>): PolygonPartsPayload {
@@ -155,14 +163,14 @@ export class HelperDB {
 
   public async find<Entity extends ObjectLiteral>(table: string, target: EntityTarget<Entity>): Promise<Entity[]> {
     const repository = this.appDataSource.getRepository(target);
-    repository.metadata.tablePath = table; // this approach may be unstable for other versions of typeorm - https://github.com/typeorm/typeorm/issues/4245#issuecomment-2134156283
+    setRepositoryTablePath(repository, table);
     const response = await repository.find();
     return response;
   }
 
   public async insert<Entity extends ObjectLiteral>(table: string, target: EntityTarget<Entity>, insertValues: Entity | Entity[]): Promise<void> {
     const repository = this.appDataSource.getRepository(target);
-    repository.metadata.tablePath = table; // this approach may be unstable for other versions of typeorm - https://github.com/typeorm/typeorm/issues/4245#issuecomment-2134156283
+    setRepositoryTablePath(repository, table);
     await repository.insert(insertValues);
   }
 }